azure-py/requirements.txt
azure-storage-blob==12.14.1